Cleaning and new feature
New feature:
before any measurement of flux or spectral binning, the data is
spectrally cropped.
Then spectral binning and existing features are applied.
Note: frame binning and creation of slices are not affected as they are
needed for this feature.

This feature duplicates the spectral binning one in the case we bin over
one pixel in the desired spectral band.