TO TEST - New fork of the extraction of flux in GLINT
This new fork address the issue of the noise in the estimation of the
flux in the photometric channels.
The idea is to stack the frames to get the spectrum.
Then, on a frame basis, we just need to integrate the bandwidth to get
the flux and reconstruct the photometric output without the detector
noise.