Changes of features
- the 'amplitude' estimator is back to an amplitude and not an integral
estimator
- method 'getSpectralFlux' takes now the absolute positions and widths
of the outputs instead of the polynomial coefficients. It takes the new
arguemt 'mode_flux' to shorten the process if 'raw' mode is selected.
- in '_getSpectralFluxNumba' and '_getSpectralFlux', linear regression
is done from the 20th column of pixel because there is noise before and
the results are absurd or the regression crash depending of the width of
the output